deduplicate tail_textfile
