+++ src/ert/scheduler/lsf_driver.py
@@ -20,6 +20,7 @@ from typing import (
     get_args,
 )
 
+from ._tail_textfile import tail_textfile
 from .driver import (
     _POLL_PERIOD,
     SIGNAL_OFFSET,
@@ -758,14 +759,3 @@ class LsfDriver(Driver):
         if msg := tail_textfile(stdout_file, num_characters_to_read_from_end):
             error_msg += f"\n    LSF-stdout:\n{msg}"
         return error_msg
-
-
-def tail_textfile(file_path: Path, num_chars: int) -> str:
-    if not file_path.exists():
-        return f"No output file {file_path}"
-    with file_path.open(encoding="utf-8") as file:
-        file.seek(0, 2)
-        file_end_position = file.tell()
-        seek_position = max(0, file_end_position - num_chars)
-        file.seek(seek_position)
-        return file.read()[-num_chars:]
