+++ src/ert/scheduler/slurm_driver.py
@@ -14,6 +14,7 @@ from enum import Enum, auto
 from pathlib import Path
 from tempfile import NamedTemporaryFile
 
+from ._tail_textfile import tail_textfile
 from .driver import (
     _POLL_PERIOD,
     SIGNAL_OFFSET,
@@ -466,23 +467,12 @@ class SlurmDriver(Driver):
     ) -> str:
         error_msg = ""
         stderr_file = Path(runpath) / (job_name + ".stderr")
-        if msg := _tail_textfile(stderr_file, num_characters_to_read_from_end):
+        if msg := tail_textfile(stderr_file, num_characters_to_read_from_end):
             error_msg += f"\n    stderr:\n{msg}"
         stdout_file = Path(runpath) / (job_name + ".stdout")
-        if msg := _tail_textfile(stdout_file, num_characters_to_read_from_end):
+        if msg := tail_textfile(stdout_file, num_characters_to_read_from_end):
             error_msg += f"\n    stdout:\n{msg}"
         return error_msg
-
-
-def _tail_textfile(file_path: Path, num_chars: int) -> str:
-    if not file_path.exists():
-        return f"No output file {file_path}"
-    with file_path.open(encoding="utf-8") as file:
-        file.seek(0, 2)
-        file_end_position = file.tell()
-        seek_position = max(0, file_end_position - num_chars)
-        file.seek(seek_position)
-        return file.read()[-num_chars:]
 
 
 def _parse_squeue_output(output: str) -> Iterator[tuple[str, SqueueInfo]]:
